+++ examples/pto/micro-op/cube-matmul/mad_bf16bf16f32/main.cpp
@@ -47,7 +47,7 @@ struct MrgSortExecutedNumList {
     }                                                                            \
   } while (0)
 
-void LaunchMad_bf16bf16f32_kernel(__bf16 *a, __bf16 *b, float *c, void *stream);
+void LaunchMad_bf16bf16f32_kernel(uint16_t *a, uint16_t *b, float *c, void *stream);
 
 int main() {
   constexpr size_t kM = 16;
@@ -57,15 +57,15 @@ int main() {
   constexpr size_t bElem = kK * kN;
   constexpr size_t cElem = kM * kN;
 
-  constexpr size_t aSize = aElem * sizeof(__bf16);
-  constexpr size_t bSize = bElem * sizeof(__bf16);
+  constexpr size_t aSize = aElem * sizeof(uint16_t);
+  constexpr size_t bSize = bElem * sizeof(uint16_t);
   constexpr size_t cSize = cElem * sizeof(float);
 
-  __bf16 *aHost = nullptr;
-  __bf16 *bHost = nullptr;
+  uint16_t *aHost = nullptr;
+  uint16_t *bHost = nullptr;
   float *cHost = nullptr;
-  __bf16 *aDevice = nullptr;
-  __bf16 *bDevice = nullptr;
+  uint16_t *aDevice = nullptr;
+  uint16_t *bDevice = nullptr;
   float *cDevice = nullptr;
 
   int rc = 0;
